factor matchers and equality implicits
